Continue the implementation of the AMQP engine

+++ src/Proton/Types/Transport/Attach.cs
@@ -344,7 +344,7 @@ namespace Apache.Qpid.Proton.Types.Transport
 
       public PerformativeType Type => PerformativeType.Attach;
 
-      public void Invoke<T>(IPerformativeHandler<T> handler, IProtonBuffer payload, int channel, T context)
+      public void Invoke<T>(IPerformativeHandler<T> handler, IProtonBuffer payload, ushort channel, T context)
       {
          handler.HandleAttach(this, payload, channel, context);
       }
